@@ -8,14 +8,12 @@ authors = [
 license = { text = "MIT" }
 readme = "README.md"
 keywords = ["Spaced-repetition"]
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
@@ -48,7 +46,7 @@ build-backend = "hatchling.build"
 
 [tool.ruff]
 # https://docs.astral.sh/ruff/
-target-version = "py39"
+target-version = "py311"
 line-length = 100
 
 [tool.ruff.lint]
@@ -81,7 +79,7 @@ known-first-party = ["spacedreppy"]
 
 [tool.mypy]
 # https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
-python_version = 3.9
+python_version = "3.11"
 pretty = true
 show_traceback = true
 color_output = true
