@@ -194,7 +194,7 @@ def test_sm2_invalid_inputs(quality, interval, repetitions, easiness):
 
 def test_spaced_repetition_scheduler_is_abstract():
     with pytest.raises(TypeError):
-        SpacedRepetitionScheduler(interval=0)
+        SpacedRepetitionScheduler(interval=0)  # type: ignore[abstract]
 
 
 def test_sm2_easiness_floor():
